Add transcript symbol column from ensembl id for human and mouse data

ensembl_to_symbol() takes as input a `tbl` formatted as | <SAMPLE> | <ENSEMBL_ID> | <COUNT> | <...> | and returns a `tbl` with the additional transcript symbol column

ensembl_to_symbol(.data, .ensembl, action = "add")

# S4 method for spec_tbl_df
ensembl_to_symbol(.data, .ensembl, action = "add")

# S4 method for tbl_df
ensembl_to_symbol(.data, .ensembl, action = "add")

# S4 method for tidybulk
ensembl_to_symbol(.data, .ensembl, action = "add")

Arguments

.data
A `tbl` formatted as | <SAMPLE> | <ENSEMBL_ID> | <COUNT> | <...> |

.ensembl
A character string. The column that is represents ensembl gene id

action
A character string. Whether to join the new information to the input tbl (add), or just get the non-redundant tbl with the new information (get).

Value

A `tbl` object including additional columns for transcript symbol

A `tbl` object including additional columns for transcript symbol

A `tbl` object including additional columns for transcript symbol

A `tbl` object including additional columns for transcript symbol

Details

This is useful since different resources use ensembl IDs while others use gene symbol IDs. At the moment this work for human (genes and transcripts) and mouse (genes) data.

Examples

  ensembl_to_symbol(tidybulk::counts_ensembl, ens)

#> # A tibble: 119 x 8
#>    ens   iso   `read count` sample cases_0_project… cases_0_samples… transcript
#>    <chr> <chr>        <dbl> <chr>  <chr>            <chr>            <chr>     
#>  1 ENSG… 13             144 TARGE… Acute Myeloid L… Primary Blood D… TSPAN6    
#>  2 ENSG… 13              72 TARGE… Acute Myeloid L… Primary Blood D… TSPAN6    
#>  3 ENSG… 13               0 TARGE… Acute Myeloid L… Primary Blood D… TSPAN6    
#>  4 ENSG… 13            1099 TARGE… Acute Myeloid L… Primary Blood D… TSPAN6    
#>  5 ENSG… 13              11 TARGE… Acute Myeloid L… Primary Blood D… TSPAN6    
#>  6 ENSG… 13               2 TARGE… Acute Myeloid L… Primary Blood D… TSPAN6    
#>  7 ENSG… 13               3 TARGE… Acute Myeloid L… Primary Blood D… TSPAN6    
#>  8 ENSG… 13            2678 TARGE… Acute Myeloid L… Primary Blood D… TSPAN6    
#>  9 ENSG… 13             751 TARGE… Acute Myeloid L… Primary Blood D… TSPAN6    
#> 10 ENSG… 13               1 TARGE… Acute Myeloid L… Primary Blood D… TSPAN6    
#> # … with 109 more rows, and 1 more variable: ref_genome <chr>