Produces the bibliography list of your workflow

get_bibliography() takes as input a `tidybulk`

get_bibliography(.data)

# S4 method for tbl
get_bibliography(.data)

# S4 method for tbl_df
get_bibliography(.data)

# S4 method for spec_tbl_df
get_bibliography(.data)

# S4 method for tidybulk
get_bibliography(.data)

# S4 method for SummarizedExperiment
get_bibliography(.data)

# S4 method for RangedSummarizedExperiment
get_bibliography(.data)

Arguments

.data

A `tbl` (with at least three columns for sample, feature and transcript abundance) or `SummarizedExperiment` (more convenient if abstracted to tibble with library(tidySummarizedExperiment))

Value

NULL. It prints a list of bibliography references for the software used through the workflow.

A consistent object (to the input) with additional columns for the statistics from the hypothesis test (e.g., log fold change, p-value and false discovery rate).

A consistent object (to the input) with additional columns for the statistics from the hypothesis test (e.g., log fold change, p-value and false discovery rate).

Details

`r lifecycle::badge("maturing")`

This methods returns the bibliography list of your workflow from the internals of a tidybulk object (attr(., "internals"))

Examples

get_bibliography(tidybulk::se_mini)
#>  @Article{tidybulk,
#>   title = {tidybulk: an R tidy framework for modular transcriptomic data analysis},
#>   author = {Stefano Mangiola and Ramyar Molania and Ruining Dong and Maria A. Doyle & Anthony T. Papenfuss},
#>   journal = {Genome Biology},
#>   year = {2021},
#>   volume = {22},
#>   number = {42},
#>   url = {https://genomebiology.biomedcentral.com/articles/10.1186/s13059-020-02233-7},
#>   }
#> @article{wickham2019welcome,
#>   title={Welcome to the Tidyverse},
#>   author={Wickham, Hadley and Averick, Mara and Bryan, Jennifer and Chang, Winston and McGowan, Lucy D'Agostino and Francois, Romain and Grolemund, Garrett and Hayes, Alex and Henry, Lionel and Hester, Jim and others},
#>   journal={Journal of Open Source Software},
#>   volume={4},
#>   number={43},
#>   pages={1686},
#>   year={2019}
#> 	}