find abundant transcripts

identify_abundant() takes as input a `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> | and returns a `tbl` with additional columns for the statistics from the hypothesis test.

identify_abundant(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  factor_of_interest = NULL,
  minimum_counts = 10,
  minimum_proportion = 0.7
)

# S4 method for spec_tbl_df
identify_abundant(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  factor_of_interest = NULL,
  minimum_counts = 10,
  minimum_proportion = 0.7
)

# S4 method for tbl_df
identify_abundant(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  factor_of_interest = NULL,
  minimum_counts = 10,
  minimum_proportion = 0.7
)

# S4 method for tidybulk
identify_abundant(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  factor_of_interest = NULL,
  minimum_counts = 10,
  minimum_proportion = 0.7
)

# S4 method for SummarizedExperiment
identify_abundant(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  factor_of_interest = NULL,
  minimum_counts = 10,
  minimum_proportion = 0.7
)

# S4 method for RangedSummarizedExperiment
identify_abundant(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  factor_of_interest = NULL,
  minimum_counts = 10,
  minimum_proportion = 0.7
)

Arguments

.data
A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |

.sample
The name of the sample column

.transcript
The name of the transcript/gene column

.abundance
The name of the transcript/gene abundance column

factor_of_interest
The name of the column of the factor of interest. This is used for defining sample groups for the filtering process. It uses the filterByExpr function from edgeR.

minimum_counts
A real positive number. It is the threshold of count per million that is used to filter transcripts/genes out from the scaling procedure.

minimum_proportion
A real positive number between 0 and 1. It is the threshold of proportion of samples for each transcripts/genes that have to be characterised by a cmp bigger than the threshold to be included for scaling procedure.

Value

A `tbl` with additional columns for the statistics from the hypothesis test (e.g., log fold change, p-value and false discovery rate).

A `tbl` with additional columns for the statistics from the hypothesis test (e.g., log fold change, p-value and false discovery rate).

A `tbl` with additional columns for the statistics from the hypothesis test (e.g., log fold change, p-value and false discovery rate).

A `tbl` with additional columns for the statistics from the hypothesis test (e.g., log fold change, p-value and false discovery rate).

A `SummarizedExperiment` object

A `SummarizedExperiment` object

Details

At the moment this function uses edgeR (DOI: 10.1093/bioinformatics/btp616)

Underlying method: edgeR::filterByExpr( data, min.count = minimum_counts, group = string_factor_of_interest, min.prop = minimum_proportion )

Examples

  identify_abundant(
  tidybulk::counts_mini,
      sample,
      transcript,
      `count`
  )

#> No group or design set. Assuming all samples belong to one group.

#> # A tibble: 2,635 x 7
#>    sample     transcript `Cell type` count time  condition .abundant
#>    <chr>      <chr>      <chr>       <dbl> <chr> <lgl>     <lgl>    
#>  1 SRR1740034 TNFRSF4    b_cell          6 0 d   TRUE      FALSE    
#>  2 SRR1740034 PLCH2      b_cell        926 0 d   TRUE      TRUE     
#>  3 SRR1740034 PADI4      b_cell         21 0 d   TRUE      TRUE     
#>  4 SRR1740034 PAX7       b_cell          0 0 d   TRUE      FALSE    
#>  5 SRR1740034 CDA        b_cell          1 0 d   TRUE      FALSE    
#>  6 SRR1740034 RCAN3      b_cell        905 0 d   TRUE      TRUE     
#>  7 SRR1740034 SMPDL3B    b_cell          3 0 d   TRUE      FALSE    
#>  8 SRR1740034 EPB41      b_cell       4667 0 d   TRUE      TRUE     
#>  9 SRR1740034 LCK        b_cell        436 0 d   TRUE      TRUE     
#> 10 SRR1740034 COL8A2     b_cell          1 0 d   TRUE      FALSE    
#> # … with 2,625 more rows