Inner join datasets

Inner join datasets

Right join datasets

Full join datasets

Arguments

x
tbls to join. (See dplyr)

y
tbls to join. (See dplyr)

by
A character vector of variables to join by. (See dplyr)

copy
If x and y are not from the same data source, and copy is TRUE, then y will be copied into the same src as x. (See dplyr)

suffix
If there are non-joined duplicate variables in x and y, these suffixes will be added to the output to disambiguate them. Should be a character vector of length 2. (See dplyr)

...
Data frames to combine (See dplyr)

Value

A tt object

A tt object

A tt object

Examples

`%>%` = magrittr::`%>%`
annotation = tidybulk::counts_SE %>% tidybulk() %>% as_tibble() %>% distinct(sample) %>% mutate(source = "AU")
tidybulk::counts_SE %>% tidybulk() %>% as_tibble() %>% inner_join(annotation)
#> Joining, by = "sample"
#> # A tibble: 408,624 × 9
#>    feature  sample count Cell.type time  condition batch factor_of_inter… source
#>    <fct>    <fct>  <dbl> <fct>     <fct> <lgl>     <fct> <lgl>            <chr> 
#>  1 A1BG     SRR17…   153 b_cell    0 d   TRUE      0     TRUE             AU    
#>  2 A1BG-AS1 SRR17…    83 b_cell    0 d   TRUE      0     TRUE             AU    
#>  3 AAAS     SRR17…   868 b_cell    0 d   TRUE      0     TRUE             AU    
#>  4 AACS     SRR17…   222 b_cell    0 d   TRUE      0     TRUE             AU    
#>  5 AAGAB    SRR17…   590 b_cell    0 d   TRUE      0     TRUE             AU    
#>  6 AAMDC    SRR17…    48 b_cell    0 d   TRUE      0     TRUE             AU    
#>  7 AAMP     SRR17…  1257 b_cell    0 d   TRUE      0     TRUE             AU    
#>  8 AANAT    SRR17…   284 b_cell    0 d   TRUE      0     TRUE             AU    
#>  9 AAR2     SRR17…   379 b_cell    0 d   TRUE      0     TRUE             AU    
#> 10 AARS2    SRR17…   685 b_cell    0 d   TRUE      0     TRUE             AU    
#> # … with 408,614 more rows

`%>%` = magrittr::`%>%`
annotation = tidybulk::counts_SE %>% tidybulk() %>% as_tibble() %>% distinct(sample) %>% mutate(source = "AU")
tidybulk::counts_SE %>% tidybulk() %>% as_tibble() %>% right_join(annotation)
#> Joining, by = "sample"
#> # A tibble: 408,624 × 9
#>    feature  sample count Cell.type time  condition batch factor_of_inter… source
#>    <fct>    <fct>  <dbl> <fct>     <fct> <lgl>     <fct> <lgl>            <chr> 
#>  1 A1BG     SRR17…   153 b_cell    0 d   TRUE      0     TRUE             AU    
#>  2 A1BG-AS1 SRR17…    83 b_cell    0 d   TRUE      0     TRUE             AU    
#>  3 AAAS     SRR17…   868 b_cell    0 d   TRUE      0     TRUE             AU    
#>  4 AACS     SRR17…   222 b_cell    0 d   TRUE      0     TRUE             AU    
#>  5 AAGAB    SRR17…   590 b_cell    0 d   TRUE      0     TRUE             AU    
#>  6 AAMDC    SRR17…    48 b_cell    0 d   TRUE      0     TRUE             AU    
#>  7 AAMP     SRR17…  1257 b_cell    0 d   TRUE      0     TRUE             AU    
#>  8 AANAT    SRR17…   284 b_cell    0 d   TRUE      0     TRUE             AU    
#>  9 AAR2     SRR17…   379 b_cell    0 d   TRUE      0     TRUE             AU    
#> 10 AARS2    SRR17…   685 b_cell    0 d   TRUE      0     TRUE             AU    
#> # … with 408,614 more rows

`%>%` = magrittr::`%>%`
annotation = tidybulk::counts_SE %>% tidybulk() %>% as_tibble() %>% distinct(sample) %>% mutate(source = "AU")
tidybulk::counts_SE %>% tidybulk() %>% as_tibble() %>% full_join(annotation)
#> Joining, by = "sample"
#> # A tibble: 408,624 × 9
#>    feature  sample count Cell.type time  condition batch factor_of_inter… source
#>    <fct>    <fct>  <dbl> <fct>     <fct> <lgl>     <fct> <lgl>            <chr> 
#>  1 A1BG     SRR17…   153 b_cell    0 d   TRUE      0     TRUE             AU    
#>  2 A1BG-AS1 SRR17…    83 b_cell    0 d   TRUE      0     TRUE             AU    
#>  3 AAAS     SRR17…   868 b_cell    0 d   TRUE      0     TRUE             AU    
#>  4 AACS     SRR17…   222 b_cell    0 d   TRUE      0     TRUE             AU    
#>  5 AAGAB    SRR17…   590 b_cell    0 d   TRUE      0     TRUE             AU    
#>  6 AAMDC    SRR17…    48 b_cell    0 d   TRUE      0     TRUE             AU    
#>  7 AAMP     SRR17…  1257 b_cell    0 d   TRUE      0     TRUE             AU    
#>  8 AANAT    SRR17…   284 b_cell    0 d   TRUE      0     TRUE             AU    
#>  9 AAR2     SRR17…   379 b_cell    0 d   TRUE      0     TRUE             AU    
#> 10 AARS2    SRR17…   685 b_cell    0 d   TRUE      0     TRUE             AU    
#> # … with 408,614 more rows