Left join datasets

Left join datasets

left_join(x, y, by = NULL, copy = FALSE, suffix = c(".x", ".y"), ...)

Arguments

x
tbls to join. (See dplyr)

y
tbls to join. (See dplyr)

by
A character vector of variables to join by. (See dplyr)

copy
If x and y are not from the same data source, and copy is TRUE, then y will be copied into the same src as x. (See dplyr)

suffix
If there are non-joined duplicate variables in x and y, these suffixes will be added to the output to disambiguate them. Should be a character vector of length 2. (See dplyr)

...
Data frames to combine (See dplyr)

Value

A tt object

Examples

`%>%` = magrittr::`%>%`
annotation = tidybulk::counts %>% distinct(sample) %>% mutate(source = "AU")
tidybulk::counts %>% left_join(annotation)

#> Joining, by = "sample"

#> # A tibble: 408,624 x 9
#>    sample transcript `Cell type` count time  condition batch factor_of_inter…
#>    <fct>  <fct>      <fct>       <dbl> <fct> <lgl>     <fct> <lgl>           
#>  1 SRR17… DDX11L1    b_cell         17 0 d   TRUE      0     TRUE            
#>  2 SRR17… WASH7P     b_cell       3568 0 d   TRUE      0     TRUE            
#>  3 SRR17… MIR6859-1  b_cell         57 0 d   TRUE      0     TRUE            
#>  4 SRR17… LOC729737  b_cell       1764 0 d   TRUE      0     TRUE            
#>  5 SRR17… MIR6859-2  b_cell         40 0 d   TRUE      0     TRUE            
#>  6 SRR17… LOC100132… b_cell       1703 0 d   TRUE      0     TRUE            
#>  7 SRR17… LOC100133… b_cell       1894 0 d   TRUE      0     TRUE            
#>  8 SRR17… LOC100288… b_cell        337 0 d   TRUE      0     TRUE            
#>  9 SRR17… FAM87B     b_cell         56 0 d   TRUE      0     TRUE            
#> 10 SRR17… LINC00115  b_cell         50 0 d   TRUE      0     TRUE            
#> # … with 408,614 more rows, and 1 more variable: source <chr>