Extract transcript-wise information

pivot_transcript() takes as input a `tbl` (with at least three columns for sample, feature and transcript abundance) or `SummarizedExperiment` (more convenient if abstracted to tibble with library(tidySummarizedExperiment)) and returns a `tbl` with only transcript-related columns

pivot_transcript(.data)

# S4 method for class 'SummarizedExperiment'
pivot_transcript(.data)

# S4 method for class 'RangedSummarizedExperiment'
pivot_transcript(.data)

Arguments

.data

A `tbl` (with at least three columns for sample, feature and transcript abundance) or `SummarizedExperiment` (more convenient if abstracted to tibble with library(tidySummarizedExperiment))

Value

A `tbl` with transcript-related information

A consistent object (to the input)

A consistent object (to the input)

Details

`r lifecycle::badge("maturing")`

This functon extracts only transcript-related information for downstream analysis (e.g., visualisation). It is disruptive in the sense that it cannot be passed anymore to tidybulk function.

References

Mangiola, S., Molania, R., Dong, R., Doyle, M. A., & Papenfuss, A. T. (2021). tidybulk: an R tidy framework for modular transcriptomic data analysis. Genome Biology, 22(1), 42. doi:10.1186/s13059-020-02233-7

Examples

## Load airway dataset for examples

  data('airway', package = 'airway')
  # Ensure a 'condition' column exists for examples expecting it

    SummarizedExperiment::colData(airway)$condition <- SummarizedExperiment::colData(airway)$dex

library(airway)
data(airway)
airway <- airway[1:100, 1:5]

  pivot_transcript(airway   )
#> # A tibble: 100 × 12
#>    .feature  gene_id gene_name entrezid gene_biotype gene_seq_start gene_seq_end
#>    <chr>     <chr>   <chr>        <int> <chr>                 <int>        <int>
#>  1 ENSG0000… ENSG00… TSPAN6          NA protein_cod…       99883667     99894988
#>  2 ENSG0000… ENSG00… TNMD            NA protein_cod…       99839799     99854882
#>  3 ENSG0000… ENSG00… DPM1            NA protein_cod…       49551404     49575092
#>  4 ENSG0000… ENSG00… SCYL3           NA protein_cod…      169818772    169863408
#>  5 ENSG0000… ENSG00… C1orf112        NA protein_cod…      169631245    169823221
#>  6 ENSG0000… ENSG00… FGR             NA protein_cod…       27938575     27961788
#>  7 ENSG0000… ENSG00… CFH             NA protein_cod…      196621008    196716634
#>  8 ENSG0000… ENSG00… FUCA2           NA protein_cod…      143815948    143832827
#>  9 ENSG0000… ENSG00… GCLC            NA protein_cod…       53362139     53481768
#> 10 ENSG0000… ENSG00… NFYA            NA protein_cod…       41040684     41067715
#> # ℹ 90 more rows
#> # ℹ 5 more variables: seq_name <chr>, seq_strand <int>, seq_coord_system <int>,
#> #   symbol <chr>, GRangesList <list>