Drop redundant elements (e.g., samples) for which feature (e.g., transcript/gene) abundances are correlated

remove_redundancy() takes as input a `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> | for correlation method or | <DIMENSION 1> | <DIMENSION 2> | <...> | for reduced_dimensions method, and returns a `tbl` with dropped elements (e.g., samples).

remove_redundancy(
  .data,
  .element = NULL,
  .feature = NULL,
  .abundance = NULL,
  method,
  of_samples = TRUE,
  correlation_threshold = 0.9,
  top = Inf,
  log_transform = FALSE,
  Dim_a_column,
  Dim_b_column
)

# S4 method for spec_tbl_df
remove_redundancy(
  .data,
  .element = NULL,
  .feature = NULL,
  .abundance = NULL,
  method,
  of_samples = TRUE,
  correlation_threshold = 0.9,
  top = Inf,
  log_transform = FALSE,
  Dim_a_column = NULL,
  Dim_b_column = NULL
)

# S4 method for tbl_df
remove_redundancy(
  .data,
  .element = NULL,
  .feature = NULL,
  .abundance = NULL,
  method,
  of_samples = TRUE,
  correlation_threshold = 0.9,
  top = Inf,
  log_transform = FALSE,
  Dim_a_column = NULL,
  Dim_b_column = NULL
)

# S4 method for tidybulk
remove_redundancy(
  .data,
  .element = NULL,
  .feature = NULL,
  .abundance = NULL,
  method,
  of_samples = TRUE,
  correlation_threshold = 0.9,
  top = Inf,
  log_transform = FALSE,
  Dim_a_column = NULL,
  Dim_b_column = NULL
)

# S4 method for SummarizedExperiment
remove_redundancy(
  .data,
  .element = NULL,
  .feature = NULL,
  .abundance = NULL,
  method,
  of_samples = TRUE,
  correlation_threshold = 0.9,
  top = Inf,
  log_transform = FALSE,
  Dim_a_column = NULL,
  Dim_b_column = NULL
)

# S4 method for RangedSummarizedExperiment
remove_redundancy(
  .data,
  .element = NULL,
  .feature = NULL,
  .abundance = NULL,
  method,
  of_samples = TRUE,
  correlation_threshold = 0.9,
  top = Inf,
  log_transform = FALSE,
  Dim_a_column = NULL,
  Dim_b_column = NULL
)

Arguments

.data
A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |

.element
The name of the element column (normally samples).

.feature
The name of the feature column (normally transcripts/genes)

.abundance
The name of the column including the numerical value the clustering is based on (normally transcript abundance)

method
A character string. The cluster algorithm to use, ay the moment k-means is the only algorithm included.

of_samples
A boolean. In case the input is a tidybulk object, it indicates Whether the element column will be sample or transcript column

correlation_threshold
A real number between 0 and 1. For correlation based calculation.

top
An integer. How many top genes to select for correlation based method

log_transform
A boolean, whether the value should be log-transformed (e.g., TRUE for RNA sequencing data)

Dim_a_column
A character string. For reduced_dimension based calculation. The column of one principal component

Dim_b_column
A character string. For reduced_dimension based calculation. The column of another principal component

Value

A tbl object with with dropped redundant elements (e.g., samples).

A tbl object with with dropped redundant elements (e.g., samples).

A tbl object with with dropped redundant elements (e.g., samples).

A tbl object with with dropped redundant elements (e.g., samples).

A `SummarizedExperiment` object

A `SummarizedExperiment` object

Details

Examples

 tidybulk::counts_mini %>% 
 tidybulk(sample, transcript, count) %>% 
 identify_abundant() %>% 
   remove_redundancy(
     .element = sample,
     .feature = transcript,
       .abundance =  count,
       method = "correlation"
       )

#> No group or design set. Assuming all samples belong to one group.

#> Getting the 527 most variable genes

#> Warning: `tbl_df()` is deprecated as of dplyr 1.0.0.
#> Please use `tibble::as_tibble()` instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

#> # A tibble: 2,108 x 7
#>    sample     transcript `Cell type` count time  condition .abundant
#>    <chr>      <chr>      <chr>       <dbl> <chr> <lgl>     <lgl>    
#>  1 SRR1740035 TNFRSF4    b_cell          6 1 d   TRUE      FALSE    
#>  2 SRR1740035 PLCH2      b_cell       1134 1 d   TRUE      TRUE     
#>  3 SRR1740035 PADI4      b_cell         17 1 d   TRUE      TRUE     
#>  4 SRR1740035 PAX7       b_cell          0 1 d   TRUE      FALSE    
#>  5 SRR1740035 CDA        b_cell          0 1 d   TRUE      FALSE    
#>  6 SRR1740035 RCAN3      b_cell        832 1 d   TRUE      TRUE     
#>  7 SRR1740035 SMPDL3B    b_cell          5 1 d   TRUE      FALSE    
#>  8 SRR1740035 EPB41      b_cell       4709 1 d   TRUE      TRUE     
#>  9 SRR1740035 LCK        b_cell        507 1 d   TRUE      TRUE     
#> 10 SRR1740035 COL8A2     b_cell          5 1 d   TRUE      FALSE    
#> # … with 2,098 more rows

counts.MDS = 
 tidybulk::counts_mini %>% 
 tidybulk(sample, transcript, count) %>% 
 identify_abundant() %>% 
  reduce_dimensions( method="MDS", .dims = 3)

#> No group or design set. Assuming all samples belong to one group.

#> tidybulk says: to access the raw results do `attr(..., "internals")$MDS`

remove_redundancy(
  counts.MDS,
  Dim_a_column = `Dim1`,
  Dim_b_column = `Dim2`,
  .element = sample,
  method = "reduced_dimensions"
)

#> # A tibble: 1,581 x 10
#>    sample transcript `Cell type` count time  condition .abundant   Dim1   Dim2
#>    <chr>  <chr>      <chr>       <dbl> <chr> <lgl>     <lgl>      <dbl>  <dbl>
#>  1 SRR17… TNFRSF4    b_cell          6 1 d   TRUE      FALSE     -0.898 -0.724
#>  2 SRR17… PLCH2      b_cell       1134 1 d   TRUE      TRUE      -0.898 -0.724
#>  3 SRR17… PADI4      b_cell         17 1 d   TRUE      TRUE      -0.898 -0.724
#>  4 SRR17… PAX7       b_cell          0 1 d   TRUE      FALSE     -0.898 -0.724
#>  5 SRR17… CDA        b_cell          0 1 d   TRUE      FALSE     -0.898 -0.724
#>  6 SRR17… RCAN3      b_cell        832 1 d   TRUE      TRUE      -0.898 -0.724
#>  7 SRR17… SMPDL3B    b_cell          5 1 d   TRUE      FALSE     -0.898 -0.724
#>  8 SRR17… EPB41      b_cell       4709 1 d   TRUE      TRUE      -0.898 -0.724
#>  9 SRR17… LCK        b_cell        507 1 d   TRUE      TRUE      -0.898 -0.724
#> 10 SRR17… COL8A2     b_cell          5 1 d   TRUE      FALSE     -0.898 -0.724
#> # … with 1,571 more rows, and 1 more variable: Dim3 <dbl>