Resolve Complete Confounders of Non-Interest

This generic function processes a SummarizedExperiment object to handle confounders that are not of interest in the analysis. It dynamically handles combinations of provided factors, adjusting the data by nesting and summarizing over these factors.

resolve_complete_confounders_of_non_interest(se, ...)

# S4 method for SummarizedExperiment
resolve_complete_confounders_of_non_interest(se, ...)

# S4 method for RangedSummarizedExperiment
resolve_complete_confounders_of_non_interest(se, ...)

Arguments

se

A SummarizedExperiment object that contains the data to be processed.

...

Arbitrary number of factor variables represented as symbols or quosures to be considered for resolving confounders. These factors are processed in combinations of two.

Value

A modified SummarizedExperiment object with confounders resolved.

A consistent object (to the input) with additional columns for the statistics from the hypothesis test (e.g., log fold change, p-value and false discovery rate).

A consistent object (to the input) with additional columns for the statistics from the hypothesis test (e.g., log fold change, p-value and false discovery rate).

Examples

# Not run:
# se is a SummarizedExperiment object
# resolve_complete_confounders_of_non_interest(se, factor1, factor2, factor3)