Add differential tissue composition information to a tbl

test_differential_cellularity() takes as input a `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> | and returns a `tbl` with additional columns for the statistics from the hypothesis test.

test_differential_cellularity(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  reference = X_cibersort,
  significance_threshold = 0.05,
  ...
)

# S4 method for spec_tbl_df
test_differential_cellularity(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  reference = X_cibersort,
  significance_threshold = 0.05,
  ...
)

# S4 method for tbl_df
test_differential_cellularity(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  reference = X_cibersort,
  significance_threshold = 0.05,
  ...
)

# S4 method for tidybulk
test_differential_cellularity(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  reference = X_cibersort,
  significance_threshold = 0.05,
  ...
)

# S4 method for SummarizedExperiment
test_differential_cellularity(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  reference = X_cibersort,
  significance_threshold = 0.05,
  ...
)

# S4 method for RangedSummarizedExperiment
test_differential_cellularity(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  reference = X_cibersort,
  significance_threshold = 0.05,
  ...
)

Arguments

.data
A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |

.formula
A formula with no response variable, representing the desired linear model. If censored regression is desired (coxph) the formula should be of the form \"survival::Surv\(y, dead\) ~ x\"

.sample
The name of the sample column

.transcript
The name of the transcript/gene column

.abundance
The name of the transcript/gene abundance column

method
A string character. Either \"cibersort\" or \"llsr\"

reference
A data frame. The transcript/cell_type data frame of integer transcript abundance

significance_threshold
A real between 0 and 1 (usually 0.05).

...
Further parameters passed to the method deconvolve_cellularity

Value

A `tbl` with additional columns for the statistics from the hypothesis test (e.g., log fold change, p-value and false discovery rate).

A `tbl` with additional columns for the statistics from the hypothesis test (e.g., log fold change, p-value and false discovery rate).

A `tbl` with additional columns for the statistics from the hypothesis test (e.g., log fold change, p-value and false discovery rate).

A `tbl` with additional columns for the statistics from the hypothesis test (e.g., log fold change, p-value and false discovery rate).

A `SummarizedExperiment` object

A `SummarizedExperiment` object

Details

This routine applies a deconvolution method (e.g., Cibersort; DOI: 10.1038/nmeth.3337) and passes the proportions inferred into a generalised linear model (DOI:dx.doi.org/10.1007/s11749-010-0189-z) or a cox regression model (ISBN: 978-1-4757-3294-8)

Underlying method for the generalised linear model: data deconvolve_cellularity( !!.sample, !!.transcript, !!.abundance, method=method, reference = reference, action="get", ... ) [..] betareg::betareg(.my_formula, .)

Underlying method for the cox regression: data deconvolve_cellularity( !!.sample, !!.transcript, !!.abundance, method=method, reference = reference, action="get", ... ) [..] mutate(.proportion_0_corrected = .proportion_0_corrected survival::coxph(.my_formula, .)

Examples

  test_differential_cellularity(
   tidybulk::counts_mini,
      ~ condition,
      sample,
      transcript,
      count,
      cores = 1
  )

#> # A tibble: 22 x 7
#>    .cell_type cell_type_propo… `estimate_(Inte… estimate_condit…
#>    <chr>      <list>                      <dbl>            <dbl>
#>  1 B cells n… <tibble [5 × 7]>            -2.55             1.74
#>  2 B cells m… <tibble [5 × 7]>            NA               NA   
#>  3 Plasma ce… <tibble [5 × 7]>            NA               NA   
#>  4 T cells C… <tibble [5 × 7]>            NA               NA   
#>  5 T cells C… <tibble [5 × 7]>            NA               NA   
#>  6 T cells C… <tibble [5 × 7]>            -3.93             2.12
#>  7 T cells C… <tibble [5 × 7]>            NA               NA   
#>  8 T cells f… <tibble [5 × 7]>            NA               NA   
#>  9 T cells r… <tibble [5 × 7]>            NA               NA   
#> 10 T cells g… <tibble [5 × 7]>            NA               NA   
#> # … with 12 more rows, and 3 more variables: std.error_conditionTRUE <dbl>,
#> #   statistic_conditionTRUE <dbl>, p.value_conditionTRUE <dbl>