analyse gene over-representation with GSEA

test_gene_overrepresentation() takes as input a `tbl` formatted as | <SAMPLE> | <ENSEMBL_ID> | <COUNT> | <...> | and returns a `tbl` with the GSEA statistics

test_gene_overrepresentation(
  .data,
  .sample = NULL,
  .entrez,
  .do_test,
  species,
  gene_set = NULL
)

# S4 method for spec_tbl_df
test_gene_overrepresentation(
  .data,
  .sample = NULL,
  .entrez,
  .do_test,
  species,
  gene_set = NULL
)

# S4 method for tbl_df
test_gene_overrepresentation(
  .data,
  .sample = NULL,
  .entrez,
  .do_test,
  species,
  gene_set = NULL
)

# S4 method for tidybulk
test_gene_overrepresentation(
  .data,
  .sample = NULL,
  .entrez,
  .do_test,
  species,
  gene_set = NULL
)

Arguments

.data
A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |

.sample
The name of the sample column

.entrez
The ENTREZ ID of the transcripts/genes

.do_test
A boolean column name symbol. It indicates the transcript to check

species
A character. For example, human or mouse. MSigDB uses the latin species names (e.g., \"Mus musculus\", \"Homo sapiens\")

gene_set
A character vector. The subset of MSigDB datasets you want to test against (e.g. \"C2\"). If NULL all gene sets are used (suggested). This argument was added to avoid time overflow of the examples.

Value

A `tbl` object

A `tbl` object

A `tbl` object

A `tbl` object

Details

This wrapper execute gene enrichment analyses of the dataset using a list of transcripts and GSEA. This wrapper uses clusterProfiler (DOI: doi.org/10.1089/omi.2011.0118) on the back-end.

Undelying method: msigdbr::msigdbr(species = species) nest(data = -gs_cat) mutate(test = map( data, ~ clusterProfiler::enricher( my_entrez_rank, TERM2GENE=.x pvalueCutoff = 1 ) ))

Examples

df_entrez = symbol_to_entrez(tidybulk::counts_mini, .transcript = transcript, .sample = sample)

#> 'select()' returned 1:many mapping between keys and columns

df_entrez = aggregate_duplicates(df_entrez, aggregation_function = sum, .sample = sample, .transcript = entrez, .abundance = count)

#> Warning: tidybulk says: for aggregation, factors and logical columns were converted to character

#> Converted to characters

#> logical

df_entrez = mutate(df_entrez, do_test = transcript %in% c("TNFRSF4", "PLCH2", "PADI4", "PAX7"))

	test_getest_gene_overrepresentation
		df_entrez,df_entrez
		.sample = sample,sample
		.entrez = entrez,entrez
		.do_test = do_test,do_test
		species="Homo sapiens""Homo sapiens"
   gene_set=c("C2")
	)

#> 

#> # A tibble: 57 x 10
#>    gs_cat ID    Description GeneRatio BgRatio  pvalue p.adjust qvalue geneID
#>    <chr>  <chr> <chr>       <chr>     <chr>     <dbl>    <dbl>  <dbl> <chr> 
#>  1 C2     HOFF… HOFFMAN_CL… 1/4       10/207… 0.00193   0.0399 0.0133 7293  
#>  2 C2     PEDR… PEDRIOLI_M… 2/4       412/20… 0.00230   0.0399 0.0133 7293/…
#>  3 C2     LIM_… LIM_MAMMAR… 1/4       14/207… 0.00270   0.0399 0.0133 9651  
#>  4 C2     MYLL… MYLLYKANGA… 1/4       18/207… 0.00347   0.0399 0.0133 5081  
#>  5 C2     GAVI… GAVIN_IL2_… 1/4       19/207… 0.00366   0.0399 0.0133 7293  
#>  6 C2     REAC… REACTOME_S… 1/4       26/207… 0.00501   0.0399 0.0133 9651  
#>  7 C2     REAC… REACTOME_T… 1/4       29/207… 0.00558   0.0399 0.0133 7293  
#>  8 C2     PEPP… PEPPER_CHR… 1/4       32/207… 0.00616   0.0399 0.0133 23569 
#>  9 C2     VALK… VALK_AML_C… 1/4       34/207… 0.00654   0.0399 0.0133 7293  
#> 10 C2     LEE_… LEE_AGING_… 1/4       37/207… 0.00712   0.0399 0.0133 7293  
#> # … with 47 more rows, and 1 more variable: Count <int>