Test of stratification of biological replicates based on tissue composition, one cell-type at the time, using Kaplan-meier curves.

test_stratification_cellularity() takes as input a `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> | and returns a `tbl` with additional columns for the statistics from the hypothesis test.

test_stratification_cellularity(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  reference = X_cibersort,
  ...
)

# S4 method for spec_tbl_df
test_stratification_cellularity(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  reference = X_cibersort,
  ...
)

# S4 method for tbl_df
test_stratification_cellularity(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  reference = X_cibersort,
  ...
)

# S4 method for tidybulk
test_stratification_cellularity(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  reference = X_cibersort,
  ...
)

Arguments

.data
A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |

.formula
A formula representing the desired linear model. The formula can be of two forms: multivariable (recommended) or univariable Respectively: \"factor_of_interest ~ .\" or \". ~ factor_of_interest\". The dot represents cell-type proportions, and it is mandatory. If censored regression is desired (coxph) the formula should be of the form \"survival::Surv\(y, dead\) ~ .\"

.sample
The name of the sample column

.transcript
The name of the transcript/gene column

.abundance
The name of the transcript/gene abundance column

method
A string character. Either \"cibersort\", \"epic\" or \"llsr\". The regression method will be chosen based on being multivariable: lm or cox-regression (both on logit-transformed proportions); or univariable: beta or cox-regression (on logit-transformed proportions). See .formula for multi- or univariable choice.

reference
A data frame. The transcript/cell_type data frame of integer transcript abundance

...
Further parameters passed to the method deconvolve_cellularity

Value

A `tbl` with additional columns for the statistics from the hypothesis test (e.g., log fold change, p-value and false discovery rate).

A `tbl` with additional columns for the statistics from the hypothesis test (e.g., log fold change, p-value and false discovery rate).

A `tbl` with additional columns for the statistics from the hypothesis test (e.g., log fold change, p-value and false discovery rate).

A `tbl` with additional columns for the statistics from the hypothesis test (e.g., log fold change, p-value and false discovery rate).

Details

This routine applies a deconvolution method (e.g., Cibersort; DOI: 10.1038/nmeth.3337) and passes the proportions inferred into a generalised linear model (DOI:dx.doi.org/10.1007/s11749-010-0189-z) or a cox regression model (ISBN: 978-1-4757-3294-8)

Underlying method for the test: data deconvolve_cellularity( !!.sample, !!.transcript, !!.abundance, method=method, reference = reference, action="get", ... ) [..] mutate(.high_cellularity = .proportion > median(.proportion)) survival::survdiff(data = data, .my_formula)

Examples

library(dplyr)
library(tidyr)

 tidybulk::counts_mini %>%

# Add survival data
nest(data = -sample) %>%
mutate(
  days = c(1, 10, 500, 1000, 2000),
  dead = c(1, 1, 1, 0, 1)
) %>%
unnest(data) %>%
test_stratification_cellularity(
  survival::Surv(days, dead) ~ .,
  sample,
  transcript,
  count,
  cores = 1
)

#> # A tibble: 22 x 6
#>    .cell_type  cell_type_propor… .low_cellularit… .high_cellulari…  pvalue plot 
#>    <chr>       <list>                       <dbl>            <dbl>   <dbl> <lis>
#>  1 B cells na… <tibble [5 × 7]>              3.35            0.65   0.0389 <ggs…
#>  2 B cells me… <tibble [5 × 7]>              3.35            0.65   0.0389 <ggs…
#>  3 Plasma cel… <tibble [5 × 7]>              2.43            1.57   0.502  <ggs…
#>  4 T cells CD8 <tibble [5 × 7]>              3.22            0.783  0.299  <ggs…
#>  5 T cells CD… <tibble [5 × 7]>              2.43            1.57   0.502  <ggs…
#>  6 T cells CD… <tibble [5 × 7]>              2.77            1.23   0.782  <ggs…
#>  7 T cells CD… <tibble [5 × 7]>              3.8             0.2    0.0455 <ggs…
#>  8 T cells fo… <tibble [5 × 7]>              3.35            0.65   0.0389 <ggs…
#>  9 T cells re… <tibble [5 × 7]>              3.22            0.783  0.774  <ggs…
#> 10 T cells ga… <tibble [5 × 7]>             NA              NA     NA      <NUL…
#> # … with 12 more rows