Creates a `tt` object from a list of file names of BAM/SAM

tidybulk_SAM_BAM() creates a `tt` object from A `tbl` (with at least three columns for sample, feature and transcript abundance) or `SummarizedExperiment` (more convenient if abstracted to tibble with library(tidySummarizedExperiment))

tidybulk_SAM_BAM(file_names, genome = "hg38", ...)

# S4 method for character,character
tidybulk_SAM_BAM(file_names, genome = "hg38", ...)

Arguments

file_names

A character vector

genome

A character string

...

Further parameters passed to the function Rsubread::featureCounts

Value

A `tidybulk` object A `tidybulk` object

Details

`r lifecycle::badge("maturing")`

This function is based on FeatureCounts package (DOI: 10.1093/bioinformatics/btt656). This function creates a tidybulk object and is useful if you want to avoid to specify .sample, .transcript and .abundance arguments all the times. The tidybulk object have an attribute called internals where these three arguments are stored as metadata. They can be extracted as attr(<object>, "internals").

Underlying core function Rsubread::featureCounts(annot.inbuilt = genome,nthreads = n_cores, ...)