Get cell type proportions from samples

deconvolve_cellularity() takes as input a `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> | and returns a `tbl` with the estimated cell type abundance for each sample

deconvolve_cellularity(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  reference = X_cibersort,
  method = "cibersort",
  action = "add",
  ...
)

# S4 method for spec_tbl_df
deconvolve_cellularity(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  reference = X_cibersort,
  method = "cibersort",
  action = "add",
  ...
)

# S4 method for tbl_df
deconvolve_cellularity(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  reference = X_cibersort,
  method = "cibersort",
  action = "add",
  ...
)

# S4 method for tidybulk
deconvolve_cellularity(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  reference = X_cibersort,
  method = "cibersort",
  action = "add",
  ...
)

# S4 method for SummarizedExperiment
deconvolve_cellularity(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  reference = X_cibersort,
  method = "cibersort",
  action = "add",
  ...
)

# S4 method for RangedSummarizedExperiment
deconvolve_cellularity(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  reference = X_cibersort,
  method = "cibersort",
  action = "add",
  ...
)

Arguments

.data
A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |

.sample
The name of the sample column

.transcript
The name of the transcript/gene column

.abundance
The name of the transcript/gene abundance column

reference
A data frame. The transcript/cell_type data frame of integer transcript abundance

method
A character string. The method to be used. At the moment Cibersort (default) and llsr (linear least squares regression) are available.

action
A character string. Whether to join the new information to the input tbl (add), or just get the non-redundant tbl with the new information (get).

...
Further parameters passed to the function Cibersort

Value

A `tbl` object including additional columns for each cell type estimated

A `tbl` object including additional columns for each cell type estimated

A `tbl` object including additional columns for each cell type estimated

A `tbl` object including additional columns for each cell type estimated

A `SummarizedExperiment` object

A `SummarizedExperiment` object

Details

This function infers the cell type composition of our samples (with the algorithm Cibersort; Newman et al., 10.1038/nmeth.3337).

Underlying method: CIBERSORT(Y = data, X = reference, ...)

Examples

# Subsetting for time efficiency
deconvolve_cellularity(filter(tidybulk::counts, sample=="SRR1740034"), sample, transcript, `count`, cores = 1)

#> # A tibble: 19,544 x 30
#>    sample transcript `Cell type` count time  condition batch factor_of_inter…
#>    <chr>  <fct>      <fct>       <dbl> <fct> <lgl>     <int> <lgl>           
#>  1 SRR17… DDX11L1    b_cell         17 0 d   TRUE          0 TRUE            
#>  2 SRR17… WASH7P     b_cell       3568 0 d   TRUE          0 TRUE            
#>  3 SRR17… MIR6859-1  b_cell         57 0 d   TRUE          0 TRUE            
#>  4 SRR17… MIR1302-2  b_cell          1 0 d   TRUE          0 TRUE            
#>  5 SRR17… FAM138A    b_cell          0 0 d   TRUE          0 TRUE            
#>  6 SRR17… OR4F5      b_cell          0 0 d   TRUE          0 TRUE            
#>  7 SRR17… LOC729737  b_cell       1764 0 d   TRUE          0 TRUE            
#>  8 SRR17… LOC102725… b_cell         11 0 d   TRUE          0 TRUE            
#>  9 SRR17… MIR6859-2  b_cell         40 0 d   TRUE          0 TRUE            
#> 10 SRR17… OR4F29     b_cell          0 0 d   TRUE          0 TRUE            
#> # … with 19,534 more rows, and 22 more variables: `cibersort: B cells
#> #   naive` <dbl>, `cibersort: B cells memory` <dbl>, `cibersort: Plasma
#> #   cells` <dbl>, `cibersort: T cells CD8` <dbl>, `cibersort: T cells CD4
#> #   naive` <dbl>, `cibersort: T cells CD4 memory resting` <dbl>, `cibersort: T
#> #   cells CD4 memory activated` <dbl>, `cibersort: T cells follicular
#> #   helper` <dbl>, `cibersort: T cells regulatory (Tregs)` <dbl>, `cibersort: T
#> #   cells gamma delta` <dbl>, `cibersort: NK cells resting` <dbl>, `cibersort:
#> #   NK cells activated` <dbl>, `cibersort: Monocytes` <dbl>, `cibersort:
#> #   Macrophages M0` <dbl>, `cibersort: Macrophages M1` <dbl>, `cibersort:
#> #   Macrophages M2` <dbl>, `cibersort: Dendritic cells resting` <dbl>,
#> #   `cibersort: Dendritic cells activated` <dbl>, `cibersort: Mast cells
#> #   resting` <dbl>, `cibersort: Mast cells activated` <dbl>, `cibersort:
#> #   Eosinophils` <dbl>, `cibersort: Neutrophils` <dbl>