Subset rows using column values

`filter()` retains the rows where the conditions you provide a `TRUE`. Note that, unlike base subsetting with `[`, rows where the condition evaluates to `NA` are dropped.

Arguments

.data
A tbl. (See dplyr)

...
<[`tidy-eval`][dplyr_tidy_eval]> Logical predicates defined in terms of the variables in `.data`. Multiple conditions are combined with `&`. Only rows where the condition evaluates to `TRUE` are kept.

.preserve
when `FALSE` (the default), the grouping structure is recalculated based on the resulting data, otherwise it is kept as is.

Value

An object of the same type as `.data`.

* Rows are a subset of the input, but appear in the same order. * Columns are not modified. * The number of groups may be reduced (if `.preserve` is not `TRUE`). * Data frame attributes are preserved.

Details

dplyr is not yet smart enough to optimise filtering optimisation on grouped datasets that don't need grouped calculations. For this reason, filtering is often considerably faster on [ungroup()]ed data.

Useful filter functions

* [`==`], [`>`], [`>=`] etc * [`&`], [`|`], [`!`], [xor()] * [is.na()] * [between()], [near()]

Grouped tibbles

Because filtering expressions are computed within groups, they may yield different results on grouped tibbles. This will be the case as soon as an aggregating, lagging, or ranking function is involved. Compare this ungrouped filtering:

The former keeps rows with `mass` greater than the global average whereas the latter keeps rows with `mass` greater than the gender

average.

Methods

This function is a **generic**, which means that packages can provide implementations (methods) for other classes. See the documentation of individual methods for extra arguments and differences in behaviour.

The following methods are currently available in loaded packages:

See also

Examples

# Learn more in ?dplyr_tidy_eval