Extract transcript-wise information

pivot_transcript() takes as input a `tbl` (with at least three columns for sample, feature and transcript abundance) or `SummarizedExperiment` (more convenient if abstracted to tibble with library(tidySummarizedExperiment)) and returns a `tbl` with only transcript-related columns

pivot_transcript(.data, .transcript = NULL)

# S4 method for class 'spec_tbl_df'
pivot_transcript(.data, .transcript = NULL)

# S4 method for class 'tbl_df'
pivot_transcript(.data, .transcript = NULL)

# S4 method for class 'tidybulk'
pivot_transcript(.data, .transcript = NULL)

# S4 method for class 'SummarizedExperiment'
pivot_transcript(.data, .transcript = NULL)

# S4 method for class 'RangedSummarizedExperiment'
pivot_transcript(.data, .transcript = NULL)

Arguments

.data

A `tbl` (with at least three columns for sample, feature and transcript abundance) or `SummarizedExperiment` (more convenient if abstracted to tibble with library(tidySummarizedExperiment))

.transcript

The name of the transcript column

Value

A `tbl` with transcript-related information

A consistent object (to the input)

A consistent object (to the input)

Details

`r lifecycle::badge("maturing")`

This functon extracts only transcript-related information for downstream analysis (e.g., visualisation). It is disruptive in the sense that it cannot be passed anymore to tidybulk function.

Examples

  pivot_transcript(tidybulk::se_mini   )
#> # A tibble: 527 × 2
#>    .feature entrez
#>    <chr>    <chr> 
#>  1 ABCB4    5244  
#>  2 ABCB9    23457 
#>  3 ACAP1    9744  
#>  4 ACHE     43    
#>  5 ACP5     54    
#>  6 ADAM28   10863 
#>  7 ADAMDEC1 27299 
#>  8 ADAMTS3  9508  
#>  9 ADRB2    154   
#> 10 AIF1     199   
#> # ℹ 517 more rows