Scale the counts of transcripts/genes

scale_abundance() takes as input a `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> | and Scales transcript abundance compansating for sequencing depth (e.g., with TMM algorithm, Robinson and Oshlack doi.org/10.1186/gb-2010-11-3-r25).

scale_abundance(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  factor_of_interest = NULL,
  minimum_counts = 10,
  minimum_proportion = 0.7,
  method = "TMM",
  reference_selection_function = median,
  action = "add"
)

# S4 method for spec_tbl_df
scale_abundance(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  factor_of_interest = NULL,
  minimum_counts = 10,
  minimum_proportion = 0.7,
  method = "TMM",
  reference_selection_function = median,
  action = "add"
)

# S4 method for tbl_df
scale_abundance(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  factor_of_interest = NULL,
  minimum_counts = 10,
  minimum_proportion = 0.7,
  method = "TMM",
  reference_selection_function = median,
  action = "add"
)

# S4 method for tidybulk
scale_abundance(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  factor_of_interest = NULL,
  minimum_counts = 10,
  minimum_proportion = 0.7,
  method = "TMM",
  reference_selection_function = median,
  action = "add"
)

# S4 method for SummarizedExperiment
scale_abundance(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  factor_of_interest = NULL,
  minimum_counts = 10,
  minimum_proportion = 0.7,
  method = "TMM",
  reference_selection_function = median,
  action = "add"
)

# S4 method for RangedSummarizedExperiment
scale_abundance(
  .data,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  factor_of_interest = NULL,
  minimum_counts = 10,
  minimum_proportion = 0.7,
  method = "TMM",
  reference_selection_function = median,
  action = "add"
)

Arguments

.data
A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |

.sample
The name of the sample column

.transcript
The name of the transcript/gene column

.abundance
The name of the transcript/gene abundance column

factor_of_interest
The name of the column of the factor of interest. This is used for identifying lowly abundant transcript, to be ignored for calculating scaling fators.

minimum_counts
A real positive number. It is the threshold of count per million that is used to filter transcripts/genes out from the scaling procedure. The scaling inference is then applied back to all unfiltered data.

minimum_proportion
A real positive number between 0 and 1. It is the threshold of proportion of samples for each transcripts/genes that have to be characterised by a cmp bigger than the threshold to be included for scaling procedure.

method
A character string. The scaling method passed to the back-end function (i.e., edgeR::calcNormFactors; "TMM","TMMwsp","RLE","upperquartile")

reference_selection_function
A function that is used to selecting the reference sample for scaling. It could be max (default), which choose the sample with maximum library size; or median, which chooses the sample with median library size.

action
A character string between "add" (default) and "only". "add" joins the new information to the input tbl (default), "only" return a non-redundant tbl with the just new information.

Value

A tbl object with additional columns with scaled data as `<NAME OF COUNT COLUMN>_scaled`

A tbl object with additional columns with scaled data as `<NAME OF COUNT COLUMN>_scaled`

A tbl object with additional columns with scaled data as `<NAME OF COUNT COLUMN>_scaled`

A tbl object with additional columns with scaled data as `<NAME OF COUNT COLUMN>_scaled`

A `SummarizedExperiment` object

A `SummarizedExperiment` object

Details

Scales transcript abundance compensating for sequencing depth (e.g., with TMM algorithm, Robinson and Oshlack doi.org/10.1186/gb-2010-11-3-r25). Lowly transcribed transcripts/genes (defined with minimum_counts and minimum_proportion parameters) are filtered out from the scaling procedure. The scaling inference is then applied back to all unfiltered data.

Underlying method edgeR::calcNormFactors(.data, method = c("TMM","TMMwsp","RLE","upperquartile"))

Examples

 scale_abundance(tidybulk::counts_mini,  sample, transcript, `count`)

#> No group or design set. Assuming all samples belong to one group.

#> # A tibble: 2,635 x 10
#>    sample transcript `Cell type` count time  condition count_scaled   TMM
#>    <chr>  <chr>      <chr>       <dbl> <chr> <lgl>            <dbl> <dbl>
#>  1 SRR17… ABCB4      b_cell       1035 0 d   TRUE           1900.   0.791
#>  2 SRR17… ABCB9      b_cell         45 0 d   TRUE             82.6  0.791
#>  3 SRR17… ACAP1      b_cell       7151 0 d   TRUE          13129.   0.791
#>  4 SRR17… ACHE       b_cell          2 0 d   TRUE              3.67 0.791
#>  5 SRR17… ACP5       b_cell       2278 0 d   TRUE           4182.   0.791
#>  6 SRR17… ADAM28     b_cell      11156 0 d   TRUE          20482.   0.791
#>  7 SRR17… ADAMDEC1   b_cell         72 0 d   TRUE            132.   0.791
#>  8 SRR17… ADAMTS3    b_cell          0 0 d   TRUE              0    0.791
#>  9 SRR17… ADRB2      b_cell        298 0 d   TRUE            547.   0.791
#> 10 SRR17… AIF1       b_cell          8 0 d   TRUE             14.7  0.791
#> # … with 2,625 more rows, and 2 more variables: multiplier <dbl>,
#> #   lowly_abundant <lgl>