Perform differential transcription testing using edgeR quasi-likelihood (QLT), edgeR likelihood-ratio (LR), limma-voom, limma-voom-with-quality-weights or DESeq2

test_differential_abundance() takes as input a `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> | and returns a `tbl` with additional columns for the statistics from the hypothesis test.

test_differential_abundance(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  .contrasts = NULL,
  method = "edgeR_quasi_likelihood",
  test_above_log2_fold_change = NULL,
  scaling_method = "TMM",
  omit_contrast_in_colnames = FALSE,
  prefix = "",
  action = "add",
  significance_threshold = NULL,
  fill_missing_values = NULL
)

# S4 method for spec_tbl_df
test_differential_abundance(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  .contrasts = NULL,
  method = "edgeR_quasi_likelihood",
  test_above_log2_fold_change = NULL,
  scaling_method = "TMM",
  omit_contrast_in_colnames = FALSE,
  prefix = "",
  action = "add",
  significance_threshold = NULL,
  fill_missing_values = NULL
)

# S4 method for tbl_df
test_differential_abundance(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  .contrasts = NULL,
  method = "edgeR_quasi_likelihood",
  test_above_log2_fold_change = NULL,
  scaling_method = "TMM",
  omit_contrast_in_colnames = FALSE,
  prefix = "",
  action = "add",
  significance_threshold = NULL,
  fill_missing_values = NULL
)

# S4 method for tidybulk
test_differential_abundance(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  .contrasts = NULL,
  method = "edgeR_quasi_likelihood",
  test_above_log2_fold_change = NULL,
  scaling_method = "TMM",
  omit_contrast_in_colnames = FALSE,
  prefix = "",
  action = "add",
  significance_threshold = NULL,
  fill_missing_values = NULL
)

# S4 method for SummarizedExperiment
test_differential_abundance(
  .data,
  .formula,
  .contrasts = NULL,
  method = "edgeR_quasi_likelihood",
  test_above_log2_fold_change = NULL,
  scaling_method = "TMM",
  omit_contrast_in_colnames = FALSE,
  prefix = ""
)

# S4 method for RangedSummarizedExperiment
test_differential_abundance(
  .data,
  .formula,
  .contrasts = NULL,
  method = "edgeR_quasi_likelihood",
  test_above_log2_fold_change = NULL,
  scaling_method = "TMM",
  omit_contrast_in_colnames = FALSE,
  prefix = ""
)

Arguments

.data
A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |

.formula
A formula representing the desired linear model. If there is more than one factor, they should be in the order factor of interest + additional factors.

.sample
The name of the sample column

.transcript
The name of the transcript/gene column

.abundance
The name of the transcript/gene abundance column

.contrasts
This parameter takes the format of the contrast parameter of the method of choice. For edgeR and limma-voom is a character vector. For DESeq2 is a list including a character vector of length three. The first covariate is the one the model is tested against (e.g., ~ factor_of_interest)

method
A string character. Either "edgeR_quasi_likelihood" (i.e., QLF), "edgeR_likelihood_ratio" (i.e., LRT), "edger_robust_likelihood_ratio", "DESeq2", "limma_voom", "limma_voom_sample_weights"

test_above_log2_fold_change
A positive real value. At the moment this works just for edgeR methods, and use the `treat` function, which test the that the difference in abundance is bigger than this parameter rather than zero https://www.rdocumentation.org/packages/edgeR/versions/3.14.0/topics/glmTreat.

scaling_method
A character string. The scaling method passed to the back-end functions: edgeR and limma-voom (i.e., edgeR::calcNormFactors; "TMM","TMMwsp","RLE","upperquartile"). Setting the parameter to \"none\" will skip the compensation for sequencing-depth for the method edgeR or limma-voom.

omit_contrast_in_colnames
If just one contrast is specified you can choose to omit the contrast label in the colnames.

prefix
A character string. The prefix you would like to add to the result columns. It is useful if you want to compare several methods.

action
A character string. Whether to join the new information to the input tbl (add), or just get the non-redundant tbl with the new information (get).

significance_threshold
DEPRECATED - A real between 0 and 1 (usually 0.05).

fill_missing_values
DEPRECATED - A boolean. Whether to fill missing sample/transcript values with the median of the transcript. This is rarely needed.

Value

A `tbl` with additional columns for the statistics from the test (e.g., log fold change, p-value and false discovery rate).

A `tbl` with additional columns for the statistics from the test (e.g., log fold change, p-value and false discovery rate).

A `tbl` with additional columns for the statistics from the hypothesis test (e.g., log fold change, p-value and false discovery rate).

A `tbl` with additional columns for the statistics from the hypothesis test (e.g., log fold change, p-value and false discovery rate).

A `SummarizedExperiment` object

A `SummarizedExperiment` object

Details

This function provides the option to use edgeR https://doi.org/10.1093/bioinformatics/btp616, limma-voom https://doi.org/10.1186/gb-2014-15-2-r29, limma_voom_sample_weights https://doi.org/10.1093/nar/gkv412 or DESeq2 https://doi.org/10.1186/s13059-014-0550-8 to perform the testing. All methods use raw counts, irrespective of if scale_abundance or adjust_abundance have been calculated, therefore it is essential to add covariates such as batch effects (if applicable) in the formula.

Underlying method for edgeR framework: .data

# Filter keep_abundant( factor_of_interest = !!(as.symbol(parse_formula(.formula)[1])), minimum_counts = minimum_counts, minimum_proportion = minimum_proportion )

# Format select(!!.transcript,!!.sample,!!.abundance) spread(!!.sample,!!.abundance) as_matrix(rownames = !!.transcript)

# edgeR edgeR::DGEList(counts = .) edgeR::calcNormFactors(method = scaling_method) edgeR::estimateDisp(design)

# Fit edgeR::glmQLFit(design) edgeR::glmQLFTest(coef = 2, contrast = my_contrasts) // or glmLRT according to choice

Underlying method for DESeq2 framework: keep_abundant( factor_of_interest = !!as.symbol(parse_formula(.formula)[[1]]), minimum_counts = minimum_counts, minimum_proportion = minimum_proportion )

# DESeq2 DESeq2::DESeqDataSet( design = .formula) DESeq2::DESeq() DESeq2::results()

Examples

 # edgeR

 tidybulk::counts_mini %>%
 tidybulk(sample, transcript, count) %>%
 identify_abundant() %>%
  test_differential_abundance( ~ condition )

#> No group or design set. Assuming all samples belong to one group.

#> =====================================
#> tidybulk says: All testing methods use raw counts, irrespective of if scale_abundance
#> or adjust_abundance have been calculated. Therefore, it is essential to add covariates
#> such as batch effects (if applicable) in the formula.
#> =====================================

#> tidybulk says: The design column names are "(Intercept), conditionTRUE"

#> tidybulk says: to access the raw results (fitted GLM) do `attr(..., "internals")$edgeR`

#> # A tibble: 2,635 x 12
#>    sample   transcript `Cell type` count time  condition .abundant  logFC logCPM
#>    <chr>    <chr>      <chr>       <dbl> <chr> <lgl>     <lgl>      <dbl>  <dbl>
#>  1 SRR1740… TNFRSF4    b_cell          6 0 d   TRUE      FALSE     NA       NA  
#>  2 SRR1740… PLCH2      b_cell        926 0 d   TRUE      TRUE       5.14    10.8
#>  3 SRR1740… PADI4      b_cell         21 0 d   TRUE      TRUE      -6.27    11.0
#>  4 SRR1740… PAX7       b_cell          0 0 d   TRUE      FALSE     NA       NA  
#>  5 SRR1740… CDA        b_cell          1 0 d   TRUE      FALSE     NA       NA  
#>  6 SRR1740… RCAN3      b_cell        905 0 d   TRUE      TRUE       2.67    11.5
#>  7 SRR1740… SMPDL3B    b_cell          3 0 d   TRUE      FALSE     NA       NA  
#>  8 SRR1740… EPB41      b_cell       4667 0 d   TRUE      TRUE       0.858   13.2
#>  9 SRR1740… LCK        b_cell        436 0 d   TRUE      TRUE       6.88    12.2
#> 10 SRR1740… COL8A2     b_cell          1 0 d   TRUE      FALSE     NA       NA  
#> # … with 2,625 more rows, and 3 more variables: F <dbl>, PValue <dbl>,
#> #   FDR <dbl>

  # The function `test_differential_abundance` operates with contrasts too

 tidybulk::counts_mini %>%
 tidybulk(sample, transcript, count) %>%
 identify_abundant() %>%
 test_differential_abundance(
      ~ 0 + condition,
      .contrasts = c( "conditionTRUE - conditionFALSE")
 )

#> No group or design set. Assuming all samples belong to one group.

#> =====================================
#> tidybulk says: All testing methods use raw counts, irrespective of if scale_abundance
#> or adjust_abundance have been calculated. Therefore, it is essential to add covariates
#> such as batch effects (if applicable) in the formula.
#> =====================================

#> tidybulk says: The design column names are "conditionFALSE, conditionTRUE"

#> tidybulk says: to access the raw results (fitted GLM) do `attr(..., "internals")$edgeR`

#> # A tibble: 2,635 x 12
#>    sample     transcript `Cell type` count time  condition .abundant
#>    <chr>      <chr>      <chr>       <dbl> <chr> <lgl>     <lgl>    
#>  1 SRR1740034 TNFRSF4    b_cell          6 0 d   TRUE      FALSE    
#>  2 SRR1740034 PLCH2      b_cell        926 0 d   TRUE      TRUE     
#>  3 SRR1740034 PADI4      b_cell         21 0 d   TRUE      TRUE     
#>  4 SRR1740034 PAX7       b_cell          0 0 d   TRUE      FALSE    
#>  5 SRR1740034 CDA        b_cell          1 0 d   TRUE      FALSE    
#>  6 SRR1740034 RCAN3      b_cell        905 0 d   TRUE      TRUE     
#>  7 SRR1740034 SMPDL3B    b_cell          3 0 d   TRUE      FALSE    
#>  8 SRR1740034 EPB41      b_cell       4667 0 d   TRUE      TRUE     
#>  9 SRR1740034 LCK        b_cell        436 0 d   TRUE      TRUE     
#> 10 SRR1740034 COL8A2     b_cell          1 0 d   TRUE      FALSE    
#> # … with 2,625 more rows, and 5 more variables:
#> #   logFC___conditionTRUE - conditionFALSE <dbl>,
#> #   logCPM___conditionTRUE - conditionFALSE <dbl>,
#> #   F___conditionTRUE - conditionFALSE <dbl>,
#> #   PValue___conditionTRUE - conditionFALSE <dbl>,
#> #   FDR___conditionTRUE - conditionFALSE <dbl>

 # DESeq2 - equivalent for limma-voom

 tidybulk::counts_mini %>%
 tidybulk(sample, transcript, count) %>%
 identify_abundant() %>%
  test_differential_abundance( ~ condition, method="deseq2" )

#> No group or design set. Assuming all samples belong to one group.

#> =====================================
#> tidybulk says: All testing methods use raw counts, irrespective of if scale_abundance
#> or adjust_abundance have been calculated. Therefore, it is essential to add covariates
#> such as batch effects (if applicable) in the formula.
#> =====================================

#> estimating size factors

#> estimating dispersions

#> gene-wise dispersion estimates

#> mean-dispersion relationship

#> -- note: fitType='parametric', but the dispersion trend was not well captured by the
#>    function: y = a/x + b, and a local regression fit was automatically substituted.
#>    specify fitType='local' or 'mean' to avoid this message next time.

#> final dispersion estimates

#> fitting model and testing

#> tidybulk says: to access the raw results (fitted GLM) do `attr(..., "internals")$DESeq2`

#> # A tibble: 2,635 x 13
#>    sample     transcript `Cell type` count time  condition .abundant baseMean
#>    <chr>      <chr>      <chr>       <dbl> <chr> <lgl>     <lgl>        <dbl>
#>  1 SRR1740034 TNFRSF4    b_cell          6 0 d   TRUE      FALSE          NA 
#>  2 SRR1740034 PLCH2      b_cell        926 0 d   TRUE      TRUE          725.
#>  3 SRR1740034 PADI4      b_cell         21 0 d   TRUE      TRUE          854.
#>  4 SRR1740034 PAX7       b_cell          0 0 d   TRUE      FALSE          NA 
#>  5 SRR1740034 CDA        b_cell          1 0 d   TRUE      FALSE          NA 
#>  6 SRR1740034 RCAN3      b_cell        905 0 d   TRUE      TRUE         1125.
#>  7 SRR1740034 SMPDL3B    b_cell          3 0 d   TRUE      FALSE          NA 
#>  8 SRR1740034 EPB41      b_cell       4667 0 d   TRUE      TRUE         3832.
#>  9 SRR1740034 LCK        b_cell        436 0 d   TRUE      TRUE         1769.
#> 10 SRR1740034 COL8A2     b_cell          1 0 d   TRUE      FALSE          NA 
#> # … with 2,625 more rows, and 5 more variables: log2FoldChange <dbl>,
#> #   lfcSE <dbl>, stat <dbl>, pvalue <dbl>, padj <dbl>

  # The function `test_differential_abundance` operates with contrasts too

 tidybulk::counts_mini %>%
 tidybulk(sample, transcript, count) %>%
 identify_abundant() %>%
 test_differential_abundance(
      ~ 0 + condition,
      .contrasts = list(c("condition", "TRUE", "FALSE")),
      method="deseq2"
 )

#> No group or design set. Assuming all samples belong to one group.

#> =====================================
#> tidybulk says: All testing methods use raw counts, irrespective of if scale_abundance
#> or adjust_abundance have been calculated. Therefore, it is essential to add covariates
#> such as batch effects (if applicable) in the formula.
#> =====================================

#> estimating size factors

#> estimating dispersions

#> gene-wise dispersion estimates

#> mean-dispersion relationship

#> -- note: fitType='parametric', but the dispersion trend was not well captured by the
#>    function: y = a/x + b, and a local regression fit was automatically substituted.
#>    specify fitType='local' or 'mean' to avoid this message next time.

#> final dispersion estimates

#> fitting model and testing

#> tidybulk says: to access the raw results (fitted GLM) do `attr(..., "internals")$DESeq2`

#> # A tibble: 2,635 x 13
#>    sample     transcript `Cell type` count time  condition .abundant
#>    <chr>      <chr>      <chr>       <dbl> <chr> <lgl>     <lgl>    
#>  1 SRR1740034 TNFRSF4    b_cell          6 0 d   TRUE      FALSE    
#>  2 SRR1740034 PLCH2      b_cell        926 0 d   TRUE      TRUE     
#>  3 SRR1740034 PADI4      b_cell         21 0 d   TRUE      TRUE     
#>  4 SRR1740034 PAX7       b_cell          0 0 d   TRUE      FALSE    
#>  5 SRR1740034 CDA        b_cell          1 0 d   TRUE      FALSE    
#>  6 SRR1740034 RCAN3      b_cell        905 0 d   TRUE      TRUE     
#>  7 SRR1740034 SMPDL3B    b_cell          3 0 d   TRUE      FALSE    
#>  8 SRR1740034 EPB41      b_cell       4667 0 d   TRUE      TRUE     
#>  9 SRR1740034 LCK        b_cell        436 0 d   TRUE      TRUE     
#> 10 SRR1740034 COL8A2     b_cell          1 0 d   TRUE      FALSE    
#> # … with 2,625 more rows, and 6 more variables:
#> #   baseMean___condition TRUE-FALSE <dbl>,
#> #   log2FoldChange___condition TRUE-FALSE <dbl>,
#> #   lfcSE___condition TRUE-FALSE <dbl>, stat___condition TRUE-FALSE <dbl>,
#> #   pvalue___condition TRUE-FALSE <dbl>, padj___condition TRUE-FALSE <dbl>