Add differential transcription information to a tbl using edgeR or DeSEQ2 from raw counts.

test_differential_abundance() takes as input a `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> | and returns a `tbl` with additional columns for the statistics from the hypothesis test.

test_differential_abundance(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  .contrasts = NULL,
  method = "edgeR_quasi_likelihood",
  significance_threshold = 0.05,
  minimum_counts = 10,
  minimum_proportion = 0.7,
  fill_missing_values = FALSE,
  scaling_method = "TMM",
  omit_contrast_in_colnames = FALSE,
  action = "add"
)

# S4 method for spec_tbl_df
test_differential_abundance(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  .contrasts = NULL,
  method = "edgeR_quasi_likelihood",
  significance_threshold = 0.05,
  minimum_counts = 10,
  minimum_proportion = 0.7,
  fill_missing_values = FALSE,
  scaling_method = "TMM",
  omit_contrast_in_colnames = FALSE,
  action = "add"
)

# S4 method for tbl_df
test_differential_abundance(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  .contrasts = NULL,
  method = "edgeR_quasi_likelihood",
  significance_threshold = 0.05,
  minimum_counts = 10,
  minimum_proportion = 0.7,
  fill_missing_values = FALSE,
  scaling_method = "TMM",
  omit_contrast_in_colnames = FALSE,
  action = "add"
)

# S4 method for tidybulk
test_differential_abundance(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  .contrasts = NULL,
  method = "edgeR_quasi_likelihood",
  significance_threshold = 0.05,
  minimum_counts = 10,
  minimum_proportion = 0.7,
  fill_missing_values = FALSE,
  scaling_method = "TMM",
  omit_contrast_in_colnames = FALSE,
  action = "add"
)

# S4 method for SummarizedExperiment
test_differential_abundance(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  .contrasts = NULL,
  method = "edgeR_quasi_likelihood",
  significance_threshold = 0.05,
  minimum_counts = 10,
  minimum_proportion = 0.7,
  fill_missing_values = FALSE,
  scaling_method = "TMM",
  omit_contrast_in_colnames = FALSE,
  action = "add"
)

# S4 method for RangedSummarizedExperiment
test_differential_abundance(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  .contrasts = NULL,
  method = "edgeR_quasi_likelihood",
  significance_threshold = 0.05,
  minimum_counts = 10,
  minimum_proportion = 0.7,
  fill_missing_values = FALSE,
  scaling_method = "TMM",
  omit_contrast_in_colnames = FALSE,
  action = "add"
)

Arguments

.data
A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |

.formula
A formula with no response variable, representing the desired linear model

.sample
The name of the sample column

.transcript
The name of the transcript/gene column

.abundance
The name of the transcript/gene abundance column

.contrasts
A character vector. See edgeR makeContrasts specification for the parameter `contrasts`. If contrasts are not present the first covariate is the one the model is tested against (e.g., ~ factor_of_interest)

method
A string character. Either "edgeR_quasi_likelihood" (i.e., QLF), "edgeR_likelihood_ratio" (i.e., LRT), "DESeq2", "limma_voom"

significance_threshold
A real between 0 and 1 (usually 0.05).

minimum_counts
A real positive number. It is the threshold of count per million that is used to filter transcripts/genes out from the scaling procedure.

minimum_proportion
A real positive number between 0 and 1. It is the threshold of proportion of samples for each transcripts/genes that have to be characterised by a cmp bigger than the threshold to be included for scaling procedure.

fill_missing_values
A boolean. Whether to fill missing sample/transcript values with the median of the transcript. This is rarely needed.

scaling_method
A character string. The scaling method passed to the back-end function (i.e., edgeR::calcNormFactors; "TMM","TMMwsp","RLE","upperquartile")

omit_contrast_in_colnames
If just one contrast is specified you can choose to omit the contrast label in the colnames.

action
A character string. Whether to join the new information to the input tbl (add), or just get the non-redundant tbl with the new information (get).

Value

A `tbl` with additional columns for the statistics from the hypothesis test (e.g., log fold change, p-value and false discovery rate).

A `tbl` with additional columns for the statistics from the hypothesis test (e.g., log fold change, p-value and false discovery rate).

A `tbl` with additional columns for the statistics from the hypothesis test (e.g., log fold change, p-value and false discovery rate).

A `tbl` with additional columns for the statistics from the hypothesis test (e.g., log fold change, p-value and false discovery rate).

A `SummarizedExperiment` object

A `SummarizedExperiment` object

Details

At the moment this function uses either edgeR (DOI: 10.18129/B9.bioc.edgeR) or DeSEQ2 (DOI: 10.1186/s13059-014-0550-8) Bith methods use raw counts, irrespectively if scale_abundance or adjust_abundance have been calculated, therefore it is essential to add covariates including unwanted source of variation in the formula.

Underlying method for edgeR framework: .data

# Filter keep_abundant( factor_of_interest = !!(as.symbol(parse_formula(.formula)[1])), minimum_counts = minimum_counts, minimum_proportion = minimum_proportion )

# Format select(!!.transcript,!!.sample,!!.abundance) spread(!!.sample,!!.abundance) as_matrix(rownames = !!.transcript)

# edgeR edgeR::DGEList(counts = .) edgeR::calcNormFactors(method = scaling_method) edgeR::estimateDisp(design)

# Fit edgeR::glmQLFit(design) edgeR::glmQLFTest(coef = 2, contrast = my_contrasts) // or glmLRT according to choice

Underlying method for DeSEQ2 framework: keep_abundant( factor_of_interest = !!as.symbol(parse_formula(.formula)[[1]]), minimum_counts = minimum_counts, minimum_proportion = minimum_proportion )

# DESeq2 DESeq2::DESeqDataSet( design = .formula) DESeq2::DESeq() DESeq2::results()

Examples

	test_ditest_differential_abundance
	 tidybultidybulks_counts_mini
	    ~ condi~icondition
	    sample,sample
	    transcrtranscript
	    `count``count`
	)

#> tidybulk says: The design column names are "(Intercept), conditionTRUE"

#> tidybulk says: to access the raw results (fitted GLM) do `attr(..., "internals")$edgeR`

#> # A tibble: 2,635 x 13
#>    sample transcript `Cell type` count time  condition  logFC logCPM     F
#>    <chr>  <chr>      <chr>       <dbl> <chr> <lgl>      <dbl>  <dbl> <dbl>
#>  1 SRR17… TNFRSF4    b_cell          6 0 d   TRUE       NA     NA    NA   
#>  2 SRR17… PLCH2      b_cell        926 0 d   TRUE        5.48  10.1  32.7 
#>  3 SRR17… PADI4      b_cell         21 0 d   TRUE       -6.08  10.1  15.3 
#>  4 SRR17… PAX7       b_cell          0 0 d   TRUE       NA     NA    NA   
#>  5 SRR17… CDA        b_cell          1 0 d   TRUE      -11.4    9.42 48.0 
#>  6 SRR17… RCAN3      b_cell        905 0 d   TRUE        2.86  10.7   5.65
#>  7 SRR17… SMPDL3B    b_cell          3 0 d   TRUE       NA     NA    NA   
#>  8 SRR17… EPB41      b_cell       4667 0 d   TRUE        1.22  12.5   2.59
#>  9 SRR17… LCK        b_cell        436 0 d   TRUE        6.99  11.3  10.1 
#> 10 SRR17… COL8A2     b_cell          1 0 d   TRUE       -6.67   7.49 36.1 
#> # … with 2,625 more rows, and 4 more variables: PValue <dbl>, FDR <dbl>,
#> #   significant <lgl>, lowly_abundant <lgl>

	# The f# The functon `test_differential_abundance` operated with contrasts too

 test_differential_abundance(
	    tidybultidybulks_counts_mini
	    ~ 0 + c~n0i+icondition
	    sample,sample
	    transcrtranscript
	    `count``count`
	    .contrasts = c( "concit"conditionTRUE - conditionFALSE"
 )

#> tidybulk says: The design column names are "conditionFALSE, conditionTRUE"

#> tidybulk says: to access the raw results (fitted GLM) do `attr(..., "internals")$edgeR`

#> # A tibble: 2,635 x 13
#>    sample transcript `Cell type` count time  condition `logFC_conditio…
#>    <chr>  <chr>      <chr>       <dbl> <chr> <lgl>                <dbl>
#>  1 SRR17… TNFRSF4    b_cell          6 0 d   TRUE                 NA   
#>  2 SRR17… PLCH2      b_cell        926 0 d   TRUE                  5.48
#>  3 SRR17… PADI4      b_cell         21 0 d   TRUE                 -6.08
#>  4 SRR17… PAX7       b_cell          0 0 d   TRUE                 NA   
#>  5 SRR17… CDA        b_cell          1 0 d   TRUE                -11.4 
#>  6 SRR17… RCAN3      b_cell        905 0 d   TRUE                  2.86
#>  7 SRR17… SMPDL3B    b_cell          3 0 d   TRUE                 NA   
#>  8 SRR17… EPB41      b_cell       4667 0 d   TRUE                  1.22
#>  9 SRR17… LCK        b_cell        436 0 d   TRUE                  6.99
#> 10 SRR17… COL8A2     b_cell          1 0 d   TRUE                 -6.67
#> # … with 2,625 more rows, and 6 more variables: `logCPM_conditionTRUE -
#> #   conditionFALSE` <dbl>, `F_conditionTRUE - conditionFALSE` <dbl>,
#> #   `PValue_conditionTRUE - conditionFALSE` <dbl>, `FDR_conditionTRUE -
#> #   conditionFALSE` <dbl>, `significant_conditionTRUE - conditionFALSE` <lgl>,
#> #   lowly_abundant <lgl>