Perform differential transcription testing using edgeR QLT, edgeR LR, limma-voom or DESeq2

test_differential_abundance() takes as input a `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> | and returns a `tbl` with additional columns for the statistics from the hypothesis test.

test_differential_abundance(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  .contrasts = NULL,
  method = "edgeR_quasi_likelihood",
  scaling_method = "TMM",
  omit_contrast_in_colnames = FALSE,
  prefix = "",
  action = "add",
  significance_threshold = NULL,
  fill_missing_values = NULL
)

# S4 method for spec_tbl_df
test_differential_abundance(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  .contrasts = NULL,
  method = "edgeR_quasi_likelihood",
  scaling_method = "TMM",
  omit_contrast_in_colnames = FALSE,
  prefix = "",
  action = "add",
  significance_threshold = NULL,
  fill_missing_values = NULL
)

# S4 method for tbl_df
test_differential_abundance(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  .contrasts = NULL,
  method = "edgeR_quasi_likelihood",
  scaling_method = "TMM",
  omit_contrast_in_colnames = FALSE,
  prefix = "",
  action = "add",
  significance_threshold = NULL,
  fill_missing_values = NULL
)

# S4 method for tidybulk
test_differential_abundance(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  .contrasts = NULL,
  method = "edgeR_quasi_likelihood",
  scaling_method = "TMM",
  omit_contrast_in_colnames = FALSE,
  prefix = "",
  action = "add",
  significance_threshold = NULL,
  fill_missing_values = NULL
)

# S4 method for SummarizedExperiment
test_differential_abundance(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  .contrasts = NULL,
  method = "edgeR_quasi_likelihood",
  scaling_method = "TMM",
  omit_contrast_in_colnames = FALSE,
  prefix = "",
  action = "add",
  significance_threshold = NULL,
  fill_missing_values = NULL
)

# S4 method for RangedSummarizedExperiment
test_differential_abundance(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  .contrasts = NULL,
  method = "edgeR_quasi_likelihood",
  scaling_method = "TMM",
  omit_contrast_in_colnames = FALSE,
  prefix = "",
  action = "add",
  significance_threshold = NULL,
  fill_missing_values = NULL
)

Arguments

.data
A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |

.formula
A formula with no response variable, representing the desired linear model

.sample
The name of the sample column

.transcript
The name of the transcript/gene column

.abundance
The name of the transcript/gene abundance column

.contrasts
A character vector. See edgeR makeContrasts specification for the parameter `contrasts`. If contrasts are not present the first covariate is the one the model is tested against (e.g., ~ factor_of_interest)

method
A string character. Either "edgeR_quasi_likelihood" (i.e., QLF), "edgeR_likelihood_ratio" (i.e., LRT), "DESeq2", "limma_voom"

scaling_method
A character string. The scaling method passed to the back-end function (i.e., edgeR::calcNormFactors; "TMM","TMMwsp","RLE","upperquartile")

omit_contrast_in_colnames
If just one contrast is specified you can choose to omit the contrast label in the colnames.

prefix
A character string. The prefix you would like to add to the result columns. It is useful if you want to compare several methods.

action
A character string. Whether to join the new information to the input tbl (add), or just get the non-redundant tbl with the new information (get).

significance_threshold
A real between 0 and 1 (usually 0.05).

fill_missing_values
A boolean. Whether to fill missing sample/transcript values with the median of the transcript. This is rarely needed.

Value

A `tbl` with additional columns for the statistics from the test (e.g., log fold change, p-value and false discovery rate).

A `tbl` with additional columns for the statistics from the test (e.g., log fold change, p-value and false discovery rate).

A `tbl` with additional columns for the statistics from the hypothesis test (e.g., log fold change, p-value and false discovery rate).

A `tbl` with additional columns for the statistics from the hypothesis test (e.g., log fold change, p-value and false discovery rate).

A `SummarizedExperiment` object

A `SummarizedExperiment` object

Details

This function provides the option to use edgeR https://doi.org/10.1093/bioinformatics/btp616, limma-voom https://doi.org/10.1186/gb-2014-15-2-r29, or DESeq2 https://doi.org/10.1186/s13059-014-0550-8 to perform the testing. All methods use raw counts, irrespective of if scale_abundance or adjust_abundance have been calculated, therefore it is essential to add covariates such as batch effects (if applicable) in the formula.

Underlying method for edgeR framework: .data

# Filter keep_abundant( factor_of_interest = !!(as.symbol(parse_formula(.formula)[1])), minimum_counts = minimum_counts, minimum_proportion = minimum_proportion )

# Format select(!!.transcript,!!.sample,!!.abundance) spread(!!.sample,!!.abundance) as_matrix(rownames = !!.transcript)

# edgeR edgeR::DGEList(counts = .) edgeR::calcNormFactors(method = scaling_method) edgeR::estimateDisp(design)

# Fit edgeR::glmQLFit(design) edgeR::glmQLFTest(coef = 2, contrast = my_contrasts) // or glmLRT according to choice

Underlying method for DESeq2 framework: keep_abundant( factor_of_interest = !!as.symbol(parse_formula(.formula)[[1]]), minimum_counts = minimum_counts, minimum_proportion = minimum_proportion )

# DESeq2 DESeq2::DESeqDataSet( design = .formula) DESeq2::DESeq() DESeq2::results()

Examples

 tidybulk::counts_mini %>% 
 tidybulk(sample, transcript, count) %>% 
 identify_abundant() %>% 
  test_differential_abundance( ~ condition )

#> No group or design set. Assuming all samples belong to one group.

#> tidybulk says: All methods use raw counts, 
#> irrespective of if scale_abundance or adjust_abundance have been calculated, 
#> therefore it is essential to add covariates such as batch effects (if applicable) in the formula.

#> tidybulk says: The design column names are "(Intercept), conditionTRUE"

#> tidybulk says: to access the raw results (fitted GLM) do `attr(..., "internals")$edgeR`

#> # A tibble: 2,635 x 12
#>    sample transcript `Cell type` count time  condition .abundant  logFC logCPM
#>    <chr>  <chr>      <chr>       <dbl> <chr> <lgl>     <lgl>      <dbl>  <dbl>
#>  1 SRR17… TNFRSF4    b_cell          6 0 d   TRUE      FALSE     NA       NA  
#>  2 SRR17… PLCH2      b_cell        926 0 d   TRUE      TRUE       5.14    10.8
#>  3 SRR17… PADI4      b_cell         21 0 d   TRUE      TRUE      -6.27    11.0
#>  4 SRR17… PAX7       b_cell          0 0 d   TRUE      FALSE     NA       NA  
#>  5 SRR17… CDA        b_cell          1 0 d   TRUE      FALSE     NA       NA  
#>  6 SRR17… RCAN3      b_cell        905 0 d   TRUE      TRUE       2.67    11.5
#>  7 SRR17… SMPDL3B    b_cell          3 0 d   TRUE      FALSE     NA       NA  
#>  8 SRR17… EPB41      b_cell       4667 0 d   TRUE      TRUE       0.858   13.2
#>  9 SRR17… LCK        b_cell        436 0 d   TRUE      TRUE       6.88    12.2
#> 10 SRR17… COL8A2     b_cell          1 0 d   TRUE      FALSE     NA       NA  
#> # … with 2,625 more rows, and 3 more variables: F <dbl>, PValue <dbl>,
#> #   FDR <dbl>

  # The function `test_differential_abundance` operates with contrasts too

 tidybulk::counts_mini %>%
 tidybulk(sample, transcript, count) %>%
 identify_abundant() %>%
 test_differential_abundance(
      ~ 0 + condition,
      .contrasts = c( "conditionTRUE - conditionFALSE")
 )

#> No group or design set. Assuming all samples belong to one group.

#> tidybulk says: All methods use raw counts, 
#> irrespective of if scale_abundance or adjust_abundance have been calculated, 
#> therefore it is essential to add covariates such as batch effects (if applicable) in the formula.

#> tidybulk says: The design column names are "conditionFALSE, conditionTRUE"

#> tidybulk says: to access the raw results (fitted GLM) do `attr(..., "internals")$edgeR`

#> # A tibble: 2,635 x 12
#>    sample transcript `Cell type` count time  condition .abundant
#>    <chr>  <chr>      <chr>       <dbl> <chr> <lgl>     <lgl>    
#>  1 SRR17… TNFRSF4    b_cell          6 0 d   TRUE      FALSE    
#>  2 SRR17… PLCH2      b_cell        926 0 d   TRUE      TRUE     
#>  3 SRR17… PADI4      b_cell         21 0 d   TRUE      TRUE     
#>  4 SRR17… PAX7       b_cell          0 0 d   TRUE      FALSE    
#>  5 SRR17… CDA        b_cell          1 0 d   TRUE      FALSE    
#>  6 SRR17… RCAN3      b_cell        905 0 d   TRUE      TRUE     
#>  7 SRR17… SMPDL3B    b_cell          3 0 d   TRUE      FALSE    
#>  8 SRR17… EPB41      b_cell       4667 0 d   TRUE      TRUE     
#>  9 SRR17… LCK        b_cell        436 0 d   TRUE      TRUE     
#> 10 SRR17… COL8A2     b_cell          1 0 d   TRUE      FALSE    
#> # … with 2,625 more rows, and 5 more variables: `logFC___conditionTRUE -
#> #   conditionFALSE` <dbl>, `logCPM___conditionTRUE - conditionFALSE` <dbl>,
#> #   `F___conditionTRUE - conditionFALSE` <dbl>, `PValue___conditionTRUE -
#> #   conditionFALSE` <dbl>, `FDR___conditionTRUE - conditionFALSE` <dbl>