Add differential tissue composition information to a tbl

test_differential_cellularity() takes as input a `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> | and returns a `tbl` with additional columns for the statistics from the hypothesis test.

test_differential_cellularity(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  reference = X_cibersort,
  significance_threshold = 0.05,
  ...
)

# S4 method for spec_tbl_df
test_differential_cellularity(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  reference = X_cibersort,
  significance_threshold = 0.05,
  ...
)

# S4 method for tbl_df
test_differential_cellularity(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  reference = X_cibersort,
  significance_threshold = 0.05,
  ...
)

# S4 method for tidybulk
test_differential_cellularity(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  reference = X_cibersort,
  significance_threshold = 0.05,
  ...
)

# S4 method for SummarizedExperiment
test_differential_cellularity(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  reference = X_cibersort,
  significance_threshold = 0.05,
  ...
)

# S4 method for RangedSummarizedExperiment
test_differential_cellularity(
  .data,
  .formula,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  method = "cibersort",
  reference = X_cibersort,
  significance_threshold = 0.05,
  ...
)

Arguments

.data
A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |

.formula
A formula representing the desired linear model. The formula can be of two forms: multivariable (recommended) or univariable Respectively: \"factor_of_interest ~ .\" or \". ~ factor_of_interest\". The dot represents cell-type proportions, and it is mandatory. If censored regression is desired (coxph) the formula should be of the form \"survival::Surv\(y, dead\) ~ .\"

.sample
The name of the sample column

.transcript
The name of the transcript/gene column

.abundance
The name of the transcript/gene abundance column

method
A string character. Either \"cibersort\", \"epic\" or \"llsr\". The regression method will be chosen based on being multivariable: lm or cox-regression (both on logit-transformed proportions); or univariable: beta or cox-regression (on logit-transformed proportions). See .formula for multi- or univariable choice.

reference
A data frame. The transcript/cell_type data frame of integer transcript abundance

significance_threshold
A real between 0 and 1 (usually 0.05).

...
Further parameters passed to the method deconvolve_cellularity

Value

A `tbl` with additional columns for the statistics from the hypothesis test (e.g., log fold change, p-value and false discovery rate).

A `tbl` with additional columns for the statistics from the hypothesis test (e.g., log fold change, p-value and false discovery rate).

A `tbl` with additional columns for the statistics from the hypothesis test (e.g., log fold change, p-value and false discovery rate).

A `tbl` with additional columns for the statistics from the hypothesis test (e.g., log fold change, p-value and false discovery rate).

A `SummarizedExperiment` object

A `SummarizedExperiment` object

Details

`r lifecycle::badge("maturing")`

This routine applies a deconvolution method (e.g., Cibersort; DOI: 10.1038/nmeth.3337) and passes the proportions inferred into a generalised linear model (DOI:dx.doi.org/10.1007/s11749-010-0189-z) or a cox regression model (ISBN: 978-1-4757-3294-8)

Underlying method for the generalised linear model: data deconvolve_cellularity( !!.sample, !!.transcript, !!.abundance, method=method, reference = reference, action="get", ... ) [..] betareg::betareg(.my_formula, .)

Underlying method for the cox regression: data deconvolve_cellularity( !!.sample, !!.transcript, !!.abundance, method=method, reference = reference, action="get", ... ) [..] mutate(.proportion_0_corrected = .proportion_0_corrected survival::coxph(.my_formula, .)

Examples

 # Regular regression
  test_differential_cellularity(
   tidybulk::se_mini ,
      . ~ condition,
      cores = 1
  )

#> # A tibble: 22 x 7
#>    .cell_type           cell_type_proport… `estimate_(Interc… estimate_conditio…
#>    <chr>                <list>                          <dbl>              <dbl>
#>  1 cibersort.B.cells.n… <tibble[,6] [5 × …              -2.55               1.74
#>  2 cibersort.B.cells.m… <tibble[,6] [5 × …              NA                 NA   
#>  3 cibersort.Plasma.ce… <tibble[,6] [5 × …              NA                 NA   
#>  4 cibersort.T.cells.C… <tibble[,6] [5 × …              NA                 NA   
#>  5 cibersort.T.cells.C… <tibble[,6] [5 × …              NA                 NA   
#>  6 cibersort.T.cells.C… <tibble[,6] [5 × …              -3.93               2.12
#>  7 cibersort.T.cells.C… <tibble[,6] [5 × …              NA                 NA   
#>  8 cibersort.T.cells.f… <tibble[,6] [5 × …              NA                 NA   
#>  9 cibersort.T.cells.r… <tibble[,6] [5 × …              NA                 NA   
#> 10 cibersort.T.cells.g… <tibble[,6] [5 × …              NA                 NA   
#> # … with 12 more rows, and 3 more variables: std.error_conditionTRUE <dbl>,
#> #   statistic_conditionTRUE <dbl>, p.value_conditionTRUE <dbl>

  # Cox regression - multiple
  library(dplyr)
  library(tidyr)

tidybulk::se_mini %>%
   tidybulk() %>%

  # Add survival data
  nest(data = -sample) %>%
  mutate(
    days = c(1, 10, 500, 1000, 2000),
    dead = c(1, 1, 1, 0, 1)
  ) %>%
  unnest(data) %>%

  # Test
  test_differential_cellularity(
      survival::Surv(days, dead) ~ .,
      cores = 1
  )

#> Warning: Ran out of iterations and did not converge

#> # A tibble: 22 x 6
#>    .cell_type           cell_type_proport… estimate std.error  statistic p.value
#>    <chr>                <list>                <dbl>     <dbl>      <dbl>   <dbl>
#>  1 B cells naive        <tibble[,7] [5 × …  112.        2205.    5.10e-2   0.959
#>  2 B cells memory       <tibble[,7] [5 × … -144.        2701.   -5.33e-2   0.958
#>  3 Plasma cells         <tibble[,7] [5 × … -320.       12172.   -2.63e-2   0.979
#>  4 T cells CD8          <tibble[,7] [5 × …    5.61      3756.    1.49e-3   0.999
#>  5 T cells CD4 naive    <tibble[,7] [5 × …    6.28      3084.    2.03e-3   0.998
#>  6 T cells CD4 memory … <tibble[,7] [5 × …    1.68      2130.    7.90e-4   0.999
#>  7 T cells CD4 memory … <tibble[,7] [5 × …    0.533    11469.    4.65e-5   1.00 
#>  8 T cells follicular … <tibble[,7] [5 × …    0         4967.    0         1    
#>  9 T cells regulatory … <tibble[,7] [5 × …   49.0       8271.    5.92e-3   0.995
#> 10 T cells gamma delta  <tibble[,7] [5 × …    0            0   NaN       NaN    
#> # … with 12 more rows