unnest

unnest

nest

unnest(
  .data,
  cols,
  ...,
  keep_empty = FALSE,
  ptype = NULL,
  names_sep = NULL,
  names_repair = "check_unique"
)

# S3 method for default
unnest(
  .data,
  cols,
  ...,
  keep_empty = FALSE,
  ptype = NULL,
  names_sep = NULL,
  names_repair = "check_unique"
)

# S3 method for nested_tidybulk
unnest(
  .data,
  cols,
  ...,
  keep_empty = FALSE,
  ptype = NULL,
  names_sep = NULL,
  names_repair = "check_unique"
)

nest(.data, ...)

# S3 method for default
nest(.data, ...)

# S3 method for tidybulk
nest(.data, ...)

Arguments

.data
A tbl. (See tidyr)

cols
<[`tidy-select`][tidyr_tidy_select]> Columns to unnest. If you `unnest()` multiple columns, parallel entries must be of compatible sizes, i.e. they're either equal or length 1 (following the standard tidyverse recycling rules).

...
Name-variable pairs of the form new_col = c(col1, col2, col3) (See tidyr)

keep_empty
See tidyr::unnest

ptype
See tidyr::unnest

names_sep
If `NULL`, the default, the names will be left as is. In `nest()`, inner names will come from the former outer names; in `unnest()`, the new outer names will come from the inner names.

If a string, the inner and outer names will be used together. In `nest()`, the names of the new outer columns will be formed by pasting together the outer and the inner column names, separated by `names_sep`. In `unnest()`, the new inner names will have the outer names (+ `names_sep`) automatically stripped. This makes `names_sep` roughly symmetric between nesting and unnesting.

names_repair
See tidyr::unnest

Value

A tt object

A tt object

Examples

library(dplyr)

nest(tidybulk(tidybulk::counts_mini, sample, transcript, count), data = -transcript) %>%
unnest(data)

#> # A tibble: 2,635 x 6
#>    transcript sample     `Cell type`       count time  condition
#>    <chr>      <chr>      <chr>             <dbl> <chr> <lgl>    
#>  1 TNFRSF4    SRR1740034 b_cell                6 0 d   TRUE     
#>  2 TNFRSF4    SRR1740035 b_cell                6 1 d   TRUE     
#>  3 TNFRSF4    SRR1740043 monocyte              1 1 d   FALSE    
#>  4 TNFRSF4    SRR1740058 t_cell              401 0 d   TRUE     
#>  5 TNFRSF4    SRR1740067 dendritic_myeloid     4 1 d   FALSE    
#>  6 PLCH2      SRR1740034 b_cell              926 0 d   TRUE     
#>  7 PLCH2      SRR1740035 b_cell             1134 1 d   TRUE     
#>  8 PLCH2      SRR1740043 monocyte             23 1 d   FALSE    
#>  9 PLCH2      SRR1740058 t_cell             2526 0 d   TRUE     
#> 10 PLCH2      SRR1740067 dendritic_myeloid    21 1 d   FALSE    
#> # … with 2,625 more rows

nest(tidybulk(tidybulk::counts_mini, sample, transcript, count), data = -transcript)

#> # A tibble: 527 x 2
#>    transcript data              
#>    <chr>      <list>            
#>  1 TNFRSF4    <tidybulk [5 × 5]>
#>  2 PLCH2      <tidybulk [5 × 5]>
#>  3 PADI4      <tidybulk [5 × 5]>
#>  4 PAX7       <tidybulk [5 × 5]>
#>  5 CDA        <tidybulk [5 × 5]>
#>  6 RCAN3      <tidybulk [5 × 5]>
#>  7 SMPDL3B    <tidybulk [5 × 5]>
#>  8 EPB41      <tidybulk [5 × 5]>
#>  9 LCK        <tidybulk [5 × 5]>
#> 10 COL8A2     <tidybulk [5 × 5]>
#> # … with 517 more rows